+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/domain/CertifiedProductSearchDetails.java
@@ -9,6 +9,7 @@ public class CertifiedProductSearchDetails {
 	
 	private Long id;
     private Long testingLabId;
+    private String testingLabName;
     private String chplProductNumber;
     private String reportFileLocation;
     private String qualityManagementSystemAtt;
@@ -26,6 +27,12 @@ public class CertifiedProductSearchDetails {
 	private Integer countCqms;
 	private Integer countCorrectiveActionPlans;
 	private Boolean visibleOnChpl;
+	private String termsOfUse;
+	private String apiDocumentation;
+	private String ics;
+	private Boolean sedTesting;
+	private Boolean qmsTesting;
+	private Boolean transparencyAttestation;
 	private Boolean privacyAttestation;
 	private Long lastModifiedDate;
 	private List<AdditionalSoftware> additionalSoftware = new ArrayList<AdditionalSoftware>();
@@ -199,5 +206,47 @@ public class CertifiedProductSearchDetails {
 	}
 	public void setCountCorrectiveActionPlans(Integer countCorrectiveActionPlans) {
 		this.countCorrectiveActionPlans = countCorrectiveActionPlans;
+	}
+	public String getTermsOfUse() {
+		return termsOfUse;
+	}
+	public void setTermsOfUse(String termsOfUse) {
+		this.termsOfUse = termsOfUse;
+	}
+	public String getApiDocumentation() {
+		return apiDocumentation;
+	}
+	public void setApiDocumentation(String apiDocumentation) {
+		this.apiDocumentation = apiDocumentation;
+	}
+	public Boolean getTransparencyAttestation() {
+		return transparencyAttestation;
+	}
+	public void setTransparencyAttestation(Boolean transparencyAttestation) {
+		this.transparencyAttestation = transparencyAttestation;
+	}
+	public String getTestingLabName() {
+		return testingLabName;
+	}
+	public void setTestingLabName(String testingLabName) {
+		this.testingLabName = testingLabName;
+	}
+	public String getIcs() {
+		return ics;
+	}
+	public void setIcs(String ics) {
+		this.ics = ics;
+	}
+	public Boolean getSedTesting() {
+		return sedTesting;
+	}
+	public void setSedTesting(Boolean sedTesting) {
+		this.sedTesting = sedTesting;
+	}
+	public Boolean getQmsTesting() {
+		return qmsTesting;
+	}
+	public void setQmsTesting(Boolean qmsTesting) {
+		this.qmsTesting = qmsTesting;
 	}
 }