+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/domain/ChplPermission.java
@@ -3,25 +3,25 @@ package gov.healthit.chpl.domain;
 import org.springframework.security.acls.domain.BasePermission;
 import org.springframework.security.acls.model.Permission;
 
-public enum CertificationBodyPermission {
+public enum ChplPermission {
 	READ(),
 	DELETE(),
 	ADMIN();
 	
-	public static Permission toPermission(CertificationBodyPermission cbp) {
+	public static Permission toPermission(ChplPermission cbp) {
 		Permission permission = null;
-		if(cbp == CertificationBodyPermission.READ) {
+		if(cbp == ChplPermission.READ) {
 			permission = BasePermission.READ;
-		} else if(cbp == CertificationBodyPermission.ADMIN) {
+		} else if(cbp == ChplPermission.ADMIN) {
 			permission = BasePermission.ADMINISTRATION;
-		} else if(cbp == CertificationBodyPermission.DELETE) {
+		} else if(cbp == ChplPermission.DELETE) {
 			permission = BasePermission.DELETE;
 		}
 		return permission;
 	}
 	
-	public static CertificationBodyPermission fromPermission(Permission perm) {
-		CertificationBodyPermission permission = null;
+	public static ChplPermission fromPermission(Permission perm) {
+		ChplPermission permission = null;
 		if(perm == BasePermission.READ) {
 			return READ;
 		} else if(perm == BasePermission.ADMINISTRATION) {