+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/dto/CertifiedProductDTO.java
@@ -27,13 +27,18 @@ public class CertifiedProductDTO {
 	private Long practiceTypeId;
 	private Long productClassificationTypeId;
 	private Long productVersionId;
-	private String qualityManagementSystemAtt;
 	private String reportFileLocation;
 	private Long testingLabId;
 	private Long certificationStatusId;
 	private String otherAcb;
 	private Boolean visibleOnChpl;
 	private Boolean privacyAttestation;
+	private String termsOfUse;
+	private String apiDocumentation;
+	private String ics;
+	private Boolean sedTesting;
+	private Boolean qmsTestig;
+	private Boolean transparencyAttestation = null;
 	
 	public CertifiedProductDTO(){}
 	
@@ -55,13 +60,17 @@ public class CertifiedProductDTO {
 		this.practiceTypeId = entity.getPracticeTypeId();
 		this.productClassificationTypeId = entity.getProductClassificationTypeId();
 		this.productVersionId = entity.getProductVersionId();
-		this.qualityManagementSystemAtt = entity.getQualityManagementSystemAtt();
 		this.reportFileLocation = entity.getReportFileLocation();
 		this.testingLabId = entity.getTestingLabId();		
 		this.certificationStatusId = entity.getCertificationStatusId();
 		this.otherAcb = entity.getOtherAcb();
 		this.setVisibleOnChpl(entity.getVisibleOnChpl());
 		this.setPrivacyAttestation(entity.getPrivacyAttestation());
+		this.setTermsOfUse(entity.getTermsOfUse());
+		this.setApiDocumentation(entity.getApiDocumentation());
+		this.setIcs(entity.getIcs());
+		this.setSedTesting(entity.getSedTesting());
+		this.setQmsTestig(entity.getQmsTesting());
 	}
 
 	
@@ -142,12 +151,6 @@ public class CertifiedProductDTO {
 	}
 	public void setProductVersionId(Long productVersionId) {
 		this.productVersionId = productVersionId;
-	}
-	public String getQualityManagementSystemAtt() {
-		return qualityManagementSystemAtt;
-	}
-	public void setQualityManagementSystemAtt(String qualityManagementSystemAtt) {
-		this.qualityManagementSystemAtt = qualityManagementSystemAtt;
 	}
 	public String getReportFileLocation() {
 		return reportFileLocation;
@@ -233,5 +236,52 @@ public class CertifiedProductDTO {
 	public void setIcsCode(String icsCode) {
 		this.icsCode = icsCode;
 	}
-	
+
+	public String getTermsOfUse() {
+		return termsOfUse;
+	}
+
+	public void setTermsOfUse(String termsOfUse) {
+		this.termsOfUse = termsOfUse;
+	}
+
+	public String getApiDocumentation() {
+		return apiDocumentation;
+	}
+
+	public void setApiDocumentation(String apiDocumentation) {
+		this.apiDocumentation = apiDocumentation;
+	}
+
+	public Boolean getTransparencyAttestation() {
+		return transparencyAttestation;
+	}
+
+	public void setTransparencyAttestation(Boolean transparencyAttestation) {
+		this.transparencyAttestation = transparencyAttestation;
+	}
+
+	public String getIcs() {
+		return ics;
+	}
+
+	public void setIcs(String ics) {
+		this.ics = ics;
+	}
+
+	public Boolean getSedTesting() {
+		return sedTesting;
+	}
+
+	public void setSedTesting(Boolean sedTesting) {
+		this.sedTesting = sedTesting;
+	}
+
+	public Boolean getQmsTestig() {
+		return qmsTestig;
+	}
+
+	public void setQmsTestig(Boolean qmsTestig) {
+		this.qmsTestig = qmsTestig;
+	}
 }