+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/web/controller/results/TestingLabResults.java
@@ -1,0 +1,22 @@
+package gov.healthit.chpl.web.controller.results;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import gov.healthit.chpl.domain.TestingLab;
+
+public class TestingLabResults {
+	private List<TestingLab> atls;
+
+	public TestingLabResults() {
+		atls = new ArrayList<TestingLab>();
+	}
+
+	public List<TestingLab> getAtls() {
+		return atls;
+	}
+
+	public void setAtls(List<TestingLab> atls) {
+		this.atls = atls;
+	}
+}