+++ RELEASE_NOTES.md
@@ -1,7 +1,21 @@
 # Release Notes
 
+## Version 0.1.0
+_5 January 2016_
+
+Features added or Updated
+* Added terms of use and api documentation to the certified product apis.
+* Added ability to update transparencyAttestation field per vendor and ACB combination. ADMINs can update the transparencyAttestation for all vendor/ACB combinations and anyone else can only update that field for the ACBs to which they have access.
+* Added transparencyAttestation to the fields that come back with certified product details and search results. It is inferred from the vendor/ACB mapping.
+* Added APIs for testing labs, found under /atls urls. 
+* Complete CHPL Product Number for 2015 products with testing lab code.
+
+Bugs Fixed
+* Correct CHPL number is inserted in corrective action plan documentation activity.
+* Authorizing existing users for new roles or ACBs/ATLs is fixed
+
 ## Version 0.0.2
-_7 December 2015__
+_7 December 2015_
 
 Features added or Updated
 * Updated the format activity streams for users are returned.